+++ src/app/admin/components/certifiedProduct/vpManagement.directive.spec.js
@@ -2,8 +2,7 @@
     'use strict';
 
     describe('chpl.admin.vpManagement.directive', function () {
-
-        var el, $log, $q, commonService, authService, vm, /*FileUploader,*/ mock, Mock, $uibModal, actualOptions;
+        var el, $log, $q, commonService, authService, vm, mock, Mock, $uibModal, actualOptions;
 
         mock = {};
         mock.developers = {developers: [{name: 'Developer 1', transparencyAttestations: []}, {name: 'Developer 2', transparencyAttestations: []}]};
@@ -40,6 +39,7 @@
                     $delegate.isAcbStaff = jasmine.createSpy('isAcbStaff');
                     $delegate.isChplAdmin = jasmine.createSpy('isChplAdmin');
                     $delegate.isOncStaff = jasmine.createSpy('isOncStaff');
+
                     return $delegate;
                 });
 
@@ -68,13 +68,9 @@
 
                     return $delegate;
                 });
-
-                $provide.decorator('FileUploader', function ($delegate) {
-                    return $delegate;
-                });
             });
 
-            inject(function ($controller, _commonService_, _authService_, /*_FileUploader_,*/ $compile, $rootScope, _$log_, _$q_, _Mock_, _$uibModal_) {
+            inject(function ($controller, _commonService_, _authService_, $compile, $rootScope, _$log_, _$q_, _Mock_, _$uibModal_) {
                 $log = _$log_;
                 $q = _$q_;
                 Mock = _Mock_;
@@ -83,8 +79,6 @@
                     actualOptions = options;
                     return Mock.fakeModal;
                 });
-                //FileUploader = _FileUploader_;
-                authService = _authService_;
                 authService = _authService_;
                 authService.getApiKey.and.returnValue($q.when('fake api key'));
                 authService.getToken.and.returnValue($q.when('fake token'));
@@ -106,7 +100,7 @@
                 commonService.getQmsStandards.and.returnValue($q.when([]));
                 commonService.getSearchOptions.and.returnValue($q.when({}));
                 commonService.getTargetedUsers.and.returnValue($q.when([]));
-                commonService.getTestFunctionality.and.returnValue($q.when([]));
+                commonService.getTestFunctionality.and.returnValue($q.when({data: []}));
                 commonService.getTestStandards.and.returnValue($q.when([]));
                 commonService.getTestTools.and.returnValue($q.when([]));
                 commonService.getUcdProcesses.and.returnValue($q.when([]));
@@ -124,7 +118,9 @@
 
         afterEach(function () {
             if ($log.debug.logs.length > 0) {
-                //console.log('\n Debug: ' + $log.debug.logs.join('\n Debug: '));
+                /* eslint-disable no-console,angular/log */
+                console.log('Debug:\n' + angular.toJson($log.debug.logs));
+                /* eslint-enable no-console,angular/log */
             }
         });
 
@@ -231,6 +227,73 @@
                 vm.splitProduct()
                 vm.splitProductInstance.dismiss('split messages');
                 expect(vm.productMessage).toBe('split messages');
+            });
+        });
+
+        describe('editing a Listing', function () {
+            var listingEditOptions;
+            beforeEach(function () {
+                listingEditOptions = {
+                    templateUrl: 'app/admin/components/certifiedProduct/certifiedProduct/edit.html',
+                    controller: 'EditCertifiedProductController',
+                    controllerAs: 'vm',
+                    animation: false,
+                    backdrop: 'static',
+                    keyboard: false,
+                    size: 'lg',
+                    resolve: {
+                        activeCP: jasmine.any(Function),
+                        isAcbAdmin: jasmine.any(Function),
+                        isAcbStaff: jasmine.any(Function),
+                        isChplAdmin: jasmine.any(Function),
+                        resources: jasmine.any(Function),
+                        workType: jasmine.any(Function)
+                    }
+                };
+            });
+
+            it('should create a modal instance when a Listing is to be edited', function () {
+                expect(vm.modalInstance).toBeUndefined();
+                vm.editCertifiedProduct()
+                expect(vm.modalInstance).toBeDefined();
+            });
+
+            it('should resolve elements on edit', function () {
+                vm.editCertifiedProduct()
+                expect($uibModal.open).toHaveBeenCalledWith(listingEditOptions);
+                expect(actualOptions.resolve.activeCP()).toEqual('');
+                actualOptions.resolve.isAcbAdmin().then(function (result) { expect(result).toEqual(true); });
+                actualOptions.resolve.isAcbStaff().then(function (result) { expect(result).toEqual(true); });
+                actualOptions.resolve.isChplAdmin().then(function (result) { expect(result).toEqual(true); });
+                expect(actualOptions.resolve.resources()).toEqual(vm.resources);
+                expect(actualOptions.resolve.workType()).toEqual(vm.workType);
+                el.isolateScope().$digest();
+            });
+
+            it('should do stuff with the returned data', function () {
+                var result = {id: 4};
+                spyOn(vm, 'refreshDevelopers');
+                spyOn(vm, 'loadCp');
+                vm.editCertifiedProduct();
+                vm.modalInstance.close(result);
+                expect(vm.activeCP).toEqual(result);
+                expect(vm.productId).toEqual(result.id);
+                expect(vm.refreshDevelopers).toHaveBeenCalled();
+                expect(vm.loadCp).toHaveBeenCalled();
+            });
+
+            it('should log a cancelled modal', function () {
+                var logCount = $log.info.logs.length;
+                vm.editCertifiedProduct();
+                vm.modalInstance.dismiss('cancelled');
+                expect($log.info.logs.length).toBe(logCount + 1);
+            });
+
+            it('should report messages if they were sent back', function () {
+                vm.editCertifiedProduct();
+                vm.modalInstance.dismiss('cancelled');
+                vm.modalInstance.dismiss('edit messages');
+                expect(vm.cpMessage).toBe('edit messages');
             });
         });
     });