+++ src/app/admin/components/surveillance/editNonconformity.controller.spec.js
@@ -36,7 +36,9 @@
 
         afterEach(function () {
             if ($log.debug.logs.length > 0) {
-                //console.debug('\n Debug: ' + $log.debug.logs.join('\n Debug: '));
+                /* eslint-disable no-console,angular/log */
+                console.debug('\n Debug: ' + $log.debug.logs.join('\n Debug: '));
+                /* eslint-enable no-console,angular/log */
             }
         });