+++ src/app/search/search.controller.js
@@ -5,7 +5,7 @@
         .controller('SearchController', SearchController);
 
     /** @ngInject */
-    function SearchController ($analytics, $filter, $localStorage, $location, $log, $rootScope, $scope, $timeout, $uibModal, cfpLoadingBar, commonService, utilService, CACHE_TIMEOUT, RELOAD_TIMEOUT) {
+    function SearchController ($analytics, $filter, $localStorage, $location, $log, $rootScope, $scope, $timeout, $uibModal, commonService, utilService, CACHE_TIMEOUT, RELOAD_TIMEOUT) {
         var vm = this;
 
         vm.browseAll = browseAll;
@@ -17,7 +17,6 @@
         vm.loadResults = loadResults;
         vm.registerAllowAll = registerAllowAll;
         vm.registerClearFilter = registerClearFilter;
-        vm.registerClearTerm = registerClearTerm;
         vm.registerRestoreComponents = registerRestoreComponents;
         vm.registerRestoreState = registerRestoreState;
         vm.registerSearch = registerSearch;
@@ -25,7 +24,6 @@
         vm.statusFont = utilService.statusFont;
         vm.triggerAllowAll = triggerAllowAll;
         vm.triggerClearFilters = triggerClearFilters;
-        vm.triggerClearTerm = triggerClearTerm;
         vm.triggerRestoreState = triggerRestoreState;
         vm.triggerSearch = triggerSearch;
         vm.viewCertificationStatusLegend = viewCertificationStatusLegend;
@@ -54,7 +52,6 @@
             vm.restoreStateHs = [];
             vm.isLoading = true;
             vm.isPreLoading = true;
-            cfpLoadingBar.start();
 
             manageStorage();
             populateSearchOptions();
@@ -93,16 +90,14 @@
         };
 
         function browseAll () {
-            $analytics.eventTrack('Browse All');
+            $analytics.eventTrack('Browse All', { category: 'Search' });
             vm.triggerClearFilters();
-            vm.triggerClearTerm();
             vm.activeSearch = true;
             setTimestamp();
         }
 
         function clear () {
             vm.triggerClearFilters();
-            vm.triggerClearTerm();
             vm.activeSearch = false;
             if (vm.searchForm) {
                 vm.searchForm.$setPristine();
@@ -139,9 +134,6 @@
 
         function loadResults() {
             commonService.getAll().then(function (response) {
-                if (vm.isPreLoading) {
-                    cfpLoadingBar.start();
-                }
                 var results = response.results;
                 for (var i = 0; i < results.length; i++) {
                     results[i].mainSearch = [results[i].developer, results[i].product, results[i].acbCertificationId, results[i].chplProductNumber].join('|');
@@ -178,16 +170,6 @@
             vm.clearFilterHs.push(handler);
             var removeHandler = function () {
                 vm.clearFilterHs = vm.clearFilterHs.filter(function (aHandler) {
-                    return aHandler !== handler;
-                });
-            };
-            return removeHandler;
-        }
-
-        function registerClearTerm (handler) {
-            vm.clearTerm = [handler]
-            var removeHandler = function () {
-                vm.clearTerm = vm.clearTerm.filter(function (aHandler) {
                     return aHandler !== handler;
                 });
             };
@@ -252,13 +234,6 @@
             vm.triggerSearch();
         }
 
-        function triggerClearTerm () {
-            angular.forEach(vm.clearTerm, function (handler) {
-                handler();
-            });
-            vm.triggerSearch();
-        }
-
         function triggerRestoreState () {
             if ($localStorage.searchTableState) {
                 var state = angular.fromJson($localStorage.searchTableState);
@@ -292,8 +267,8 @@
         }
 
         function viewPreviouslyCompared (doNotSearch) {
-            $analytics.eventTrack('View Previously Compared');
             if (!doNotSearch) {
+                vm.triggerClearFilters();
                 vm.triggerAllowAll();
             }
             $localStorage.viewingPreviouslyCompared = true;
@@ -303,13 +278,14 @@
                 vm.previouslyIds.push({value: id, selected: true})
             });
             if (!doNotSearch) {
+                $analytics.eventTrack('View Previously Compared', { category: 'Search' });
                 vm.triggerSearch();
             }
         }
 
         function viewPreviouslyViewed (doNotSearch) {
-            $analytics.eventTrack('View Previously Viewed');
             if (!doNotSearch) {
+                vm.triggerClearFilters();
                 vm.triggerAllowAll();
             }
             $localStorage.viewingPreviouslyViewed = true;
@@ -319,6 +295,7 @@
                 vm.previouslyIds.push({value: id, selected: true})
             });
             if (!doNotSearch) {
+                $analytics.eventTrack('View Previously Viewed', { category: 'Search' });
                 vm.triggerSearch();
             }
         }
@@ -380,10 +357,6 @@
             vm.lookaheadSource = {all: [], developers: [], products: []};
             commonService.getSearchOptions(true)
                 .then(function (options) {
-                    if (vm.isPreLoading) {
-                        cfpLoadingBar.start();
-                    }
-
                     vm.searchOptions = options;
                     var i;
                     options.practiceTypes = [];
@@ -424,11 +397,9 @@
                 if (difference > CACHE_TIMEOUT) {
                     vm.activeSearch = false;
                 } else {
-                    cfpLoadingBar.start();
                     $timeout(
                         function () {
                             vm.triggerRestoreState();
-                            cfpLoadingBar.complete();
                         },
                         RELOAD_TIMEOUT
                     );