+++ RELEASE_NOTES.md
@@ -1,5 +1,25 @@
 # Release Notes
 
+## Version 8.3.0
+_8 May 2017_
+
+### New Features
+* Allow optionally banning Developers on appropriate status change
+* Tweaked GTM events and expanded the events that are tracked
+* Show description instead of just name on Functionality Tested & Standard
+* Changed description of "Non-conformities" file on download page
+* Allow Privacy & Security value: `Approach 1;Approach 2`
+* Update Activity reports for greater usability
+* Added section for Notifications: Surveillance Activity, though currently hidden
+* Displayed filtered Test Functionalities on Criteria editing
+
+### Bugs fixed
+* Fixed bug related to Merged Developer status history
+* Fixed bug related to too long Name/Additional Software in CMS PDF generation
+* Viewing previously viewed/compared clears all filters to allow it
+
+---
+
 ## Version 8.2.0
 _24 April 2017_